feat: Hex-counter support for orders

+++ order/core/src/main/kotlin/com/rarible/protocol/order/core/model/SeaportOrderData.kt
@@ -28,7 +28,7 @@ data class SeaportOrderComponents(
     override val zoneHash: Word,
     override val salt: BigInteger,
     override val conduitKey: Word,
-    val counter: Long
+    val counter: BigInteger
 ) : SeaportOrderData()
 
 data class SeaportOrderParameters(
